ci: the two Windows legs' first run with the demo found a deploy race and two findings
Both Windows legs failed on the first run that built the demo, for three
reasons, none of them the demo's code.

Four windeployqt passes share bin/: ac3gui's and ac3gui_qmltests's
(ordered, apps/gui/tests says why), ac3desk's and ac3desk_qmltests's
(ordered likewise). The GUI's pair and the demo's pair were not ordered
against each other, and two of them copying the same Qt file at once fail
with "Cannot create a file when that file already exists" - the MSVC leg
lost ac3gui's copy, the clang-cl leg lost ac3desk's. The demo's pair now
chains after whichever of the GUI's is last, so the four run one at a time.

version.cpp's new headline branch tests a constexpr count, which MSVC at
/W4 reports as a constant condition (C4127) on the arm64 leg; it is now
if constexpr, which is what it always was.

The runner passed a float level through printf's varargs, which clang-cl
reports as a promotion to double (-Wdouble-promotion); it is cast to say so.

+++ apps/windows/runner/main.cpp
@@ -79,7 +79,7 @@ void print_apps(const ac3::windemo::EngineStatus& s) {
                                 std::to_string(app.position.z).substr(0, 5) + ")")
                                    .c_str()
                              : "bed",
-                    app.fullscreen ? " [full-screen]" : "", app.level_dbfs);
+                    app.fullscreen ? " [full-screen]" : "", static_cast<double>(app.level_dbfs));
         std::putchar('\n');
     }
 }
